fix: move extraIntegration to integration.yaml (#2467)

+++ jobs/post-install-job/job/migrations/integration-type/git_parser.go
@@ -19,12 +19,27 @@ type manifest struct {
 	IntegrationType integration.Type `json:"IntegrationType" yaml:"IntegrationType"`
 }
 
-type ExtraIntegrations struct {
-	URLs []string `json:"extraIntegrations" yaml:"extraIntegrations"`
+type IntegrationYaml struct {
+	Plugins []struct {
+		ID              int                 `json:"id" yaml:"id"`
+		IntegrationType integration.Type    `json:"integration_type" yaml:"integration_type"`
+		Name            string              `json:"name" yaml:"name"`
+		Tier            string              `json:"tier" yaml:"tier"`
+		Tags            map[string][]string `json:"tags" yaml:"tags"`
+		Description     string              `json:"description" yaml:"description"`
+		Icon            string              `json:"icon" yaml:"icon"`
+		Availability    string              `json:"availability" yaml:"availability"`
+		SourceCode      string              `json:"source_code" yaml:"source_code"`
+		PackageType     string              `json:"package_type" yaml:"package_type"`
+		ArtifactDetails struct {
+			PackageURL string `json:"package_url" yaml:"package_url"`
+			PackageTag string `json:"package_tag" yaml:"package_tag"`
+		} `json:"artifact_details" yaml:"artifact_details"`
+	} `json:"plugins" yaml:"plugins"`
 }
 
 func (g *GitParser) ExtractIntegrationBinaries(logger *zap.Logger) error {
-	var extraIntegrations ExtraIntegrations
+	var integrationYaml IntegrationYaml
 	// read file from path
 	f, err := os.Open(config.IntegrationTypesYamlPath)
 	if err != nil {
@@ -32,7 +47,7 @@ func (g *GitParser) ExtractIntegrationBinaries(logger *zap.Logger) error {
 		return fmt.Errorf("open file: %w", err)
 	}
 	defer f.Close()
-	if err := yaml.NewDecoder(f).Decode(&extraIntegrations); err != nil {
+	if err := yaml.NewDecoder(f).Decode(&integrationYaml); err != nil {
 		logger.Error("failed to decode json", zap.Error(err))
 		return fmt.Errorf("decode json: %w", err)
 	}
@@ -48,11 +63,15 @@ func (g *GitParser) ExtractIntegrationBinaries(logger *zap.Logger) error {
 	}
 
 	// download files from urls
-	for _, url := range extraIntegrations.URLs {
+	for _, iPlugin := range integrationYaml.Plugins {
+		if iPlugin.ArtifactDetails.PackageURL == "" || iPlugin.ArtifactDetails.PackageTag != "" {
+			continue
+		}
+		url := iPlugin.ArtifactDetails.PackageURL
 		// remove existing files
 		if err := os.RemoveAll(baseDir + "/integarion_type"); err != nil {
 			logger.Error("failed to remove existing files", zap.Error(err), zap.String("url", url), zap.String("path", baseDir+"/integarion_type"))
-			return fmt.Errorf("remove existing files for url %s: %w", url, err)
+			return fmt.Errorf("remove existing files for url %s: %w", iPlugin, err)
 		}
 
 		downloader := getter.Client{
@@ -63,38 +82,38 @@ func (g *GitParser) ExtractIntegrationBinaries(logger *zap.Logger) error {
 		err := downloader.Get()
 		if err != nil {
 			logger.Error("failed to get integration binaries", zap.Error(err), zap.String("url", url))
-			return fmt.Errorf("get integration binaries for url %s: %w", url, err)
+			return fmt.Errorf("get integration binaries for url %s: %w", iPlugin, err)
 		}
 
-		// read manifest file
-		manifestFile, err := os.Open(baseDir + "/integarion_type/manifest.yaml")
-		if err != nil {
-			logger.Error("failed to open manifest file", zap.Error(err))
-			return fmt.Errorf("open manifest file: %w", err)
-		}
-		defer manifestFile.Close()
-		var m manifest
-		// decode yaml
-		if err := yaml.NewDecoder(manifestFile).Decode(&m); err != nil {
-			logger.Error("failed to decode manifest", zap.Error(err), zap.String("url", url))
-			return fmt.Errorf("decode manifest for url %s: %w", url, err)
-		}
+		//// read manifest file
+		//manifestFile, err := os.Open(baseDir + "/integarion_type/manifest.yaml")
+		//if err != nil {
+		//	logger.Error("failed to open manifest file", zap.Error(err))
+		//	return fmt.Errorf("open manifest file: %w", err)
+		//}
+		//defer manifestFile.Close()
+		//var m manifest
+		//// decode yaml
+		//if err := yaml.NewDecoder(manifestFile).Decode(&m); err != nil {
+		//	logger.Error("failed to decode manifest", zap.Error(err), zap.String("url", url))
+		//	return fmt.Errorf("decode manifest for url %s: %w", iPlugin, err)
+		//}
 
 		// read integration-plugin file
 		integrationPlugin, err := os.ReadFile(baseDir + "/integarion_type/integration-plugin")
 		if err != nil {
 			logger.Error("failed to open integration-plugin file", zap.Error(err), zap.String("url", url))
-			return fmt.Errorf("open integration-plugin file for url %s: %w", url, err)
+			return fmt.Errorf("open integration-plugin file for url %s: %w", iPlugin, err)
 		}
 		cloudqlPlugin, err := os.ReadFile(baseDir + "/integarion_type/cloudql-plugin")
 		if err != nil {
 			logger.Error("failed to open cloudql-plugin file", zap.Error(err), zap.String("url", url))
-			return fmt.Errorf("open cloudql-plugin file for url %s: %w", url, err)
+			return fmt.Errorf("open cloudql-plugin file for url %s: %w", iPlugin, err)
 		}
 
-		logger.Info("done reading files", zap.String("url", url), zap.String("integrationType", m.IntegrationType.String()), zap.Int("integrationPluginSize", len(integrationPlugin)), zap.Int("cloudqlPluginSize", len(cloudqlPlugin)))
+		logger.Info("done reading files", zap.String("url", url), zap.String("integrationType", iPlugin.IntegrationType.String()), zap.Int("integrationPluginSize", len(integrationPlugin)), zap.Int("cloudqlPluginSize", len(cloudqlPlugin)))
 		g.IntegrationBinaries = append(g.IntegrationBinaries, models.IntegrationTypeBinaries{
-			IntegrationType:   m.IntegrationType,
+			IntegrationType:   iPlugin.IntegrationType,
 			IntegrationPlugin: integrationPlugin,
 			CloudQlPlugin:     cloudqlPlugin,
 		})
